Added dependencies and things for deployment
Please approve immediately.
