+++ .env
@@ -1,0 +1,3 @@
+PORT=3000
+MONGODB_URL="mongodb+srv://axel:axel123@obscuracluster-2swgt.mongodb.net/obscura?retryWrites=true&w=majority"
+SESSION_SECRET="somegibberishsecret"
