+++ config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  envPort: process.env.PORT,
+  dbURL: process.env.MONGODB_URL,
+  sessionKey: process.env.SESSION_SECRET,
+};
