+++ models/connection.js
@@ -1,0 +1,12 @@
+const mongoose = require('mongoose');
+const { dbURL } = require('../config');
+
+const options = { 
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+  useFindAndModify: false 
+};
+
+mongoose.connect(dbURL, options);
+
+module.exports = mongoose;
